add admin http handlermanager
